pygments_data: Move data to typescript.
