+++ web/src/pygments_data.ts
@@ -1,0 +1,6 @@
+import generated_pygments_data from "../generated/pygments_data.json";
+
+type PygmentsLanguage = {priority: number; pretty_name: string};
+const langs: Record<string, PygmentsLanguage | undefined> = generated_pygments_data.langs;
+
+export {langs};
