+++ web/src/ui_init.js
@@ -2,7 +2,6 @@ import $ from "jquery";
 import _ from "lodash";
 
 import generated_emoji_codes from "../../static/generated/emoji/emoji_codes.json";
-import generated_pygments_data from "../generated/pygments_data.json";
 import * as fenced_code from "../shared/src/fenced_code";
 import render_compose from "../templates/compose.hbs";
 import render_message_feed_bottom_whitespace from "../templates/message_feed_bottom_whitespace.hbs";
@@ -87,6 +86,7 @@ import * as pm_list from "./pm_list";
 import * as popover_menus from "./popover_menus";
 import * as popovers from "./popovers";
 import * as presence from "./presence";
+import * as pygments_data from "./pygments_data";
 import * as realm_logo from "./realm_logo";
 import * as realm_playground from "./realm_playground";
 import * as realm_user_settings_defaults from "./realm_user_settings_defaults";
@@ -699,7 +699,7 @@ export function initialize_everything() {
     typing.initialize();
     starred_messages_ui.initialize();
     user_status_ui.initialize();
-    fenced_code.initialize(generated_pygments_data);
+    fenced_code.initialize(pygments_data);
     message_edit_history.initialize();
     hotkey.initialize();
     desktop_integration.initialize();
