+++ web/tests/typeahead_helper.test.js
@@ -18,7 +18,7 @@ const stream_data = zrequire("stream_data");
 const stream_list_sort = zrequire("stream_list_sort");
 const compose_state = zrequire("compose_state");
 const emoji = zrequire("emoji");
-const pygments_data = zrequire("../generated/pygments_data.json");
+const pygments_data = zrequire("pygments_data");
 const util = zrequire("util");
 const actual_pygments_data = {...pygments_data};
 const ct = zrequire("composebox_typeahead");
